Revert "Fixed up first draft engine"
This reverts commit 8fe4223c8ada98dcf1a266eddc1ea0b691bf4a6e.

+++ tw-project/src/trig/game/entity/dummy/DummyBullet.java
@@ -18,7 +18,7 @@ import java.util.Random;
  */
 public final class DummyBullet extends Projectile implements Visible
 {
-    protected Shape shape = new RoundRectangle2D.Float(hitSize, hitSize, hitSize, hitSize, (float) 0, (float) 0);
+    protected Shape shape = new RoundRectangle2D.Float(hitSize, hitSize, hitSize, hitSize, 0, 0);
 
     @Override
     public Shape getShape(){
